Move Preference Dialogs to same package
