+++ app/src/main/scala/com/waz/zclient/preferences/dialogs/PingIntervalPref.scala
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.waz.zclient.preferences
+package com.waz.zclient.preferences.dialogs
 
 import android.content.Context
 import android.support.v7.preference.EditTextPreference
